make doveauth also use generic dictproxy
